+++ R/plausible_values.R
@@ -14,7 +14,9 @@
 #' A covariate must be a discrete person property (e.g. not a float) that indicates nominal categories, e.g. gender or school.
 #' If dataSrc is a data.frame, it must contain the covariate.
 #' @param nPV Number of plausible values to draw per person.
-#' @param use_draw When the ENORM was fitted with a Gibbs sampler, this specifies the use of a particular sample of item parameters used to generate the plausible value(s). If NULL, the posterior means are used. If outside range, the last iteration will be used.
+#' @param use_draw When the ENORM was fitted with a Gibbs sampler, 
+#' this specifies the use of a particular sample of item parameters used to generate the plausible value(s). 
+#' If NULL, the posterior means are used. 
 #' @param prior.dist use a normal prior or a mixture of two normals
 #' @param merge_within_persons If a person took multiple booklets, this indicates
 #' whether plausible values are generated per person (TRUE) or per booklet (FALSE)